modularized Inspect
