+++ app/containers/Inspect/styles.scss
