--- app/stylesheets/Applications.scss
@@ -1,3 +1,0 @@
-@import './constants.scss';
-
-
