--- app/stylesheets/Contact.scss
@@ -1,167 +1,0 @@
-@import './constants.scss';
-
-.contact {
-  display: flex;
-  flex-direction: column;
-  flex-direction: column;
-  align-content: center;
-  align-items: center;
-  justify-content: center;
-  justify-items: center;
-  top: 0;
-  left: 0;
-  min-height: 100vh;
-  opacity: 1;
-  vertical-align: top;
-  justify-content: center;
-  align-items: center;
-  color: $background;
-}
-.contact-border {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  font-size: 18px;
-  border-radius: 10px;
-  border: none;
-  box-shadow: $boxshadow2;
-  padding: 30px;
-  min-width: 421px;
-  max-width: 600px;
-  margin: 100px 20px;
-
-  &:hover {
-    background-color: rgba(255, 255, 255, 0.1);
-  }
-}
-.contact-container {
-  padding: 20px;
-  margin-right: 0;
-  justify-content: left;
-  min-width: 50%;
-}
-.contact-blurb {
-  min-width: 50%;
-  font-size: 18px;
-}
-.email-container {
-  line-height: 34px;
-  font-size: 18px;
-  text-align: start;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-content: flex-start;
-  align-items: flex-start;
-  align-self: flex-start;
-  padding: 0px 20px;
-  margin: 0;
-  width: 100%;
-  form {
-    width: 100%;
-    label {
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-      input {
-        border: 1px solid transparent;
-        border-radius: 5px;
-        width: 100%;
-        color: $header;
-        font-size: 16px;
-        padding: 6px;
-        &[type='file'] {
-          width: 226px;
-          cursor: pointer;
-        }
-      }
-      textarea {
-        border-radius: 5px;
-        font-size: 16px;
-        width: 100%;
-        min-width: 100%;
-        max-width: 100%;
-        height: 150px;
-        min-height: 50px;
-        max-height: 300px;
-        padding: 6px;
-      }
-    }
-  }
-
-  #messageLabel {
-    vertical-align: top;
-  }
-
-  #message {
-    margin-top: 10px;
-    border-color: lightgray;
-    color: $header;
-  }
-}
-
-#myfile {
-  font-size: 16px;
-  color: $background;
-}
-
-#contact-submit {
-  border: none;
-  color: $background;
-
-  &:hover {
-    border: none;
-    cursor: pointer;
-  }
-
-  &:focus {
-    border: none;
-  }
-
-  &:active {
-    border: none;
-  }
-}
-
-input[type='file'] {
-  border: none;
-  border-radius: 5px;
-
-  &:hover {
-    border: none;
-  }
-
-  &:focus {
-    border: none;
-  }
-
-  &:active {
-    border: none;
-  }
-}
-
-#contact-submit {
-  font-size: 18px;
-  width: 90px;
-  height: 40px;
-  margin-top: 20px;
-  margin-left: -4px;
-  padding: 10px;
-  border: 1px dotted $background;
-  outline: none;
-  color: $background;
-  background-color: transparent;
-  border-radius: 8px;
-  &:hover {
-    background-color: $gblue;
-  }
-}
-
-#issueLink {
-  color: $header;
-  cursor: pointer;
-  text-decoration: underline;
-  &:hover {
-    color: blue !important;
-  }
-}
