--- app/stylesheets/ServicesList.scss
@@ -1,23 +1,0 @@
-@import './constants.scss';
-
-/* .services-btn {
-  font-family: 'Nunito', sans-serif;
-  background-color: #24262F;
-  width: 13em;
-  color: white;
-  word-wrap: break-word;
-  font-size: 15px;
-  text-align: center;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-  border: none;
-  height: 50px;
-  transition: 0.2s;
-  margin: 0 8px 0 8px;
-}
-
-.services-btn:hover {
-  background-color: #494FEE;
-  font-weight: 400;
-  box-shadow: #333;
-  cursor: pointer;
-} */
